OneMedia/requirements.txt
# Tlg parse and bot libs
joblib==1.0.1
nltk==3.6.2
psycopg2-binary==2.8.6
Telethon==1.21.1
urllib3==1.26.4
cachetools==4.2.4
# vk
vk_api==11.9.6
# Web Crawlers libs
aiohttp[speedups]==3.7.4.post0
beautifulsoup4==4.9.3
aiohttp-socks==0.6.0
python-socks==1.2.4
stem==1.8.0
fake-useragent==0.1.11
asyncpg==0.23.0
dateparser==1.0.0
openpyxl==3.0.9
lxml==4.7.1
feedparser==6.0.8
readability-lxml==0.8.1
newspaper3k==0.2.8
# monitoring to check cpu usage
psutil==5.8.0
PyJWT==1.7.1
regex==2022.3.2